Fixed all the bugs for 2 nodes (3 nodes still bugged)

+++ Chord/src/FixFingers.java
@@ -22,7 +22,7 @@ public class FixFingers extends Thread {
 				for (int next = 0; next<32; next++) {
 					Thread.sleep(sleepTimeMillis);
 					FingerObject currentFingerEntry = node.getFingerTable().get(next);
-					Long currentId = node.getNodeUpperBound() + (long) (Math.pow(2, next) % Math.pow(2, 32));
+					Long currentId = (node.getNodeUpperBound() + (long) (Math.pow(2, next))) % (long) Math.pow(2, 32);
 					currentFingerEntry.setAddress(node.findSuccessor(currentId));
 				}
 			} catch (InterruptedException e) {
